+++ example/lib/route.dart
@@ -2,15 +2,18 @@ import 'home.dart';
 import 'package:flutter/widgets.dart';
 
 import 'simple.dart';
+import 'test_appear_condition.dart';
+import 'test_page_condition.dart';
+import 'test_position_condition.dart';
 
 class SectionViewRoute {
   static const String initialRoute = "/";
   static final Map<String, WidgetBuilder> routes = {
     "/": (context) => Stack(
           children: const [
-            HomePage(
-              title: "Home",
-            ),
+            // TestPositionConditionPage(),
+            HomePage()
+            // SimplePage()
           ],
         ),
   };